feat: adding game code on class menu that runs all screens
